# contracts/rest/openapi.yaml
openapi: 3.1.0
info:
  title: Arca Bench - Test & Evaluation Laboratory
  version: 2.3.0
  description: Synchronous contract of arca-bench, generated from the running FastAPI
    app (see ADR-002, ADR-009).
paths:
  /api/v1/bench/health:
    get:
      tags:
      - arca-bench
      summary: Api Health
      operationId: api_health_api_v1_bench_health_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /api/v1/tests:
    get:
      tags:
      - arca-bench
      summary: List Tests
      operationId: list_tests_api_v1_tests_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
    post:
      tags:
      - arca-bench
      summary: Register Test
      operationId: register_test_api_v1_tests_post
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: true
              type: object
              title: Body
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runs:
    get:
      tags:
      - arca-bench
      summary: List Runs
      operationId: list_runs_api_v1_runs_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
    post:
      tags:
      - arca-bench
      summary: Start Run
      operationId: start_run_api_v1_runs_post
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: true
              type: object
              title: Body
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runs/{run_id}:
    get:
      tags:
      - arca-bench
      summary: Get Run
      operationId: get_run_api_v1_runs__run_id__get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runs/{run_id}/results:
    post:
      tags:
      - arca-bench
      summary: Submit Result
      operationId: submit_result_api_v1_runs__run_id__results_post
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              title: Body
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runs/{run_id}/complete:
    post:
      tags:
      - arca-bench
      summary: Complete Run
      operationId: complete_run_api_v1_runs__run_id__complete_post
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runs/{run_id}/evidence:
    post:
      tags:
      - arca-bench
      summary: Collect Evidence
      operationId: collect_evidence_api_v1_runs__run_id__evidence_post
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              title: Body
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - arca-bench
      summary: List Evidence
      operationId: list_evidence_api_v1_runs__run_id__evidence_get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/runs/{run_id}/report:
    get:
      tags:
      - arca-bench
      summary: Get Report
      operationId: get_report_api_v1_runs__run_id__report_get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      - name: format
        in: query
        required: false
        schema:
          type: string
          default: json
          title: Format
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/scores:
    get:
      tags:
      - arca-bench
      summary: List Scores
      operationId: list_scores_api_v1_scores_get
      parameters:
      - name: target
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Target
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/scorecards/{run_id}:
    get:
      tags:
      - arca-bench
      summary: Get Scorecard
      operationId: get_scorecard_api_v1_scorecards__run_id__get
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          title: Run Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/decision-replay/rules:
    get:
      tags:
      - arca-bench
      summary: List Replay Rules
      operationId: list_replay_rules_api_v1_decision_replay_rules_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /api/v1/decision-replay/evaluate:
    post:
      tags:
      - arca-bench
      summary: Evaluate Decision
      operationId: evaluate_decision_api_v1_decision_replay_evaluate_post
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: true
              type: object
              title: Body
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /mcp/capabilities:
    get:
      tags:
      - mcp
      summary: List Capabilities
      operationId: list_capabilities_mcp_capabilities_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /mcp/tools/test-register:
    post:
      tags:
      - mcp
      summary: Test Register
      description: Register a new test case via MCP.
      operationId: test_register_mcp_tools_test_register_post
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: true
              type: object
              title: Body
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /mcp/tools/run_test_suite:
    post:
      tags:
      - mcp
      summary: Run Test Suite
      operationId: run_test_suite_mcp_tools_run_test_suite_post
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: true
              type: object
              title: Body
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /mcp/tools/get_scorecard:
    post:
      tags:
      - mcp
      summary: Get Scorecard
      operationId: get_scorecard_mcp_tools_get_scorecard_post
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: true
              type: object
              title: Body
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /mcp/tools/generate_report:
    post:
      tags:
      - mcp
      summary: Generate Report
      operationId: generate_report_mcp_tools_generate_report_post
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: true
              type: object
              title: Body
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /graphql:
    get:
      summary: Handle Http Get
      operationId: handle_http_get_graphql_get
      responses:
        '200':
          description: The GraphiQL integrated development environment.
          content:
            application/json:
              schema: {}
        '404':
          description: Not found if GraphiQL or query via GET are not enabled.
    post:
      summary: Handle Http Post
      operationId: handle_http_post_graphql_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /healthz:
    get:
      summary: Healthz
      operationId: healthz_healthz_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /readyz:
    get:
      summary: Readyz
      operationId: readyz_readyz_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError